Split sonatype module
Changes:
* split `sonatype` module in 3: `sonatype-central`, `sonatype-cp` and `sonatype-nx2`
* the `sonatype-central` defines central requirements **only**
* the `sonatype-cp` has the `sonatype-cp` publisher and support classes and deps
* the `sonatype-nx2` has the Nx2 support classes and `sonatype-oss` and `sonatype-s01` publishers
* the `apache` publisher now depends on `sonatype-nx2` only

All of these depends on `sonatype-central` as they all publish to central.
