+++ pom.xml
@@ -107,7 +107,17 @@
       </dependency>
       <dependency>
         <groupId>eu.maveniverse.maven.njord.publisher</groupId>
-        <artifactId>sonatype</artifactId>
+        <artifactId>sonatype-central</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>eu.maveniverse.maven.njord.publisher</groupId>
+        <artifactId>sonatype-cp</artifactId>
+        <version>${project.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>eu.maveniverse.maven.njord.publisher</groupId>
+        <artifactId>sonatype-nx2</artifactId>
         <version>${project.version}</version>
       </dependency>
