+++ publisher/sonatype-central/README.md
@@ -1,0 +1,3 @@
+# Sonatype publisher support
+
+This module does not implements any publishers. It implements Central Requirements only.
