+++ publisher/sonatype-central/pom.xml
@@ -17,7 +17,7 @@
     <version>0.8.3-SNAPSHOT</version>
   </parent>
 
-  <artifactId>apache</artifactId>
+  <artifactId>sonatype-central</artifactId>
 
   <name>${project.groupId}:${project.artifactId}</name>
 
@@ -25,14 +25,6 @@
     <dependency>
       <groupId>eu.maveniverse.maven.njord</groupId>
       <artifactId>core</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>eu.maveniverse.maven.njord.publisher</groupId>
-      <artifactId>sonatype-central</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>eu.maveniverse.maven.njord.publisher</groupId>
-      <artifactId>sonatype-nx2</artifactId>
     </dependency>
 
     <dependency>
